pyproject.toml
[tool.poetry]
name = "diploma-smart-parking"
version = "0.1.0"
description = ""
authors = ["Dzhurunov Tamerlan <dzhurunovt@gmail.com>"]
readme = "README.md"
packages = [{include = "diploma_smart_parking", from = "src"}]

[tool.poetry.dependencies]
python = ">=3.11"
django = "^5.1.6"
djangorestframework = "^3.15.2"
gunicorn = "^23.0.0"
psycopg2-binary = "^2.9.10"
django-cors-headers = "^4.7.0"
djangorestframework-simplejwt = "^5.3.1"
pillow = "^11.2.1"

[tool.poetry.group.dev.dependencies]
pytest-django = "^4.8.0"
pytest-cov = "^4.1.0"
pytest = "^8.3.5"
typing-extensions = "^4.8.0"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
